model_w_rag/requirements.txt
langchain==0.2.14
langchain-community==0.2.12
langchain-openai==0.1.23
chromadb>=0.5.0
sentence-transformers>=2.3.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
setuptools
